Conflicts:
	build.conf
	casa5/gcwrap/tasks/imview.xml
	casa5/gcwrap/tasks/msview.xml

+++ casatools/src/cmpt/iterbot/iterbotsink_cmpt.cc
@@ -63,7 +63,7 @@ namespace casac {
         return fromRecord(state.getSummaryRecord( ));
     }
 
-    int iterbotsink::cleanComplete(const bool lastcyclecheck) {
+    long iterbotsink::cleanComplete(const bool lastcyclecheck) {
         return state.cleanComplete(lastcyclecheck);
     }
